Aw yeah. I added Obj loading and drawing! Now I just need to figure out why animations are playing. Maybe it has something to do with me not resolving UOL's yet...

+++ NoLifeStory/Time.cpp
@@ -13,7 +13,7 @@ void NLS::_Time::Reset() {
 }
 
 void NLS::_Time::Step() {
-	double temp = clock.GetElapsedTime();
+	uint32_t temp = clock.GetElapsedTime();
 	delta = temp-tdelta;
 	tdelta = temp;
 }
